fix ui

+++ app/models/Product.scala
@@ -4,7 +4,7 @@ case class Product(
                     private val name: String,
                     private val category: String,
                     private val description: String,
-                    private val price: Float
+                    private val price: BigDecimal
                   ) {
 
   def getName = name
